express37-game/bot/requirements.txt
python-telegram-bot==20.7
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0